requirements.txt
Django==1.7.4
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-immutablefield==0.1.1
django-toolbelt==0.0.1
djangorestframework==3.0.5
gunicorn==19.2.1
mock==1.0.1
psycopg2==2.6
static3==0.5.1
wsgiref==0.1.2
